feat: Add map functionality to sidebar and header
