+++ src/app/page.tsx
@@ -1,5 +1,4 @@
 "use client";
-import { useState, useEffect } from "react";
 import { Card } from "~/components/card/";
 import { SearchInput } from "~/components/search-input";
 import { api } from "~/trpc/react";
@@ -9,42 +8,13 @@ import { Skeleton } from "~/components/ui/skeleton";
 import { useDebouncedState } from "~/hooks/use-debouced-state";
 import { Alert, AlertDescription, AlertTitle } from "~/components/ui/alert";
 import { FiInfo } from "react-icons/fi";
-import { MapContainer, TileLayer, Marker, Popup, useMap } from "react-leaflet";
 import "leaflet/dist/leaflet.css";
 import "leaflet-defaulticon-compatibility/dist/leaflet-defaulticon-compatibility.css";
 import "leaflet-defaulticon-compatibility";
-import { type LatLngTuple } from "leaflet";
-
-function UserLocationMap({ userLocation }: { userLocation: LatLngTuple }) {
-  const map = useMap();
-
-  useEffect(() => {
-    if (userLocation) {
-      map.setView(userLocation, 13);
-    }
-  }, [userLocation, map]);
-
-  return null;
-}
 
 export default function Home() {
   const { data, isLoading } = api.shelter.findAll.useQuery();
   const [searchTerm, setSearchTerm] = useDebouncedState("", 300);
-  const [userLocation, setUserLocation] = useState<LatLngTuple>([
-    -30.0346, -51.2177,
-  ]); // Default to Porto Alegre
-
-  useEffect(() => {
-    navigator.geolocation.getCurrentPosition(
-      (position) => {
-        const { latitude, longitude } = position.coords;
-        setUserLocation([latitude, longitude]);
-      },
-      (error) => {
-        console.error("Error getting location: ", error);
-      },
-    );
-  }, []);
 
   const filteredShelters = useMemo(() => {
     const trimmedSearchTerm = searchTerm.trim();
@@ -74,41 +44,6 @@ export default function Home() {
 
   return (
     <main className="flex w-full flex-col items-center justify-center gap-2 bg-white px-3 pt-8">
-      <MapContainer
-        style={{ height: "400px", width: "100%", maxWidth: "1280px" }}
-        center={userLocation}
-        zoom={13}
-      >
-        <UserLocationMap userLocation={userLocation} />
-        <TileLayer
-          attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
-          url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-        />
-        {data?.map((shelter) => {
-          if (!shelter.latitude || !shelter.longitude) return null;
-
-          return (
-            <Marker
-              key={shelter.id}
-              position={[shelter.latitude, shelter.longitude]}
-            >
-              <Popup>
-                <div className="flex flex-col space-y-2">
-                  <h2 className="text-lg font-bold">{shelter.name}</h2>
-                  <p>{shelter.phone}</p>
-                  <p>{shelter.addressStreet}</p>
-                  <p>
-                    {shelter.addressCity} - {shelter.addressState}
-                  </p>
-                  <p>{shelter.addressNeighborhood}</p>
-                  <p>{shelter.addressZip}</p>
-                </div>
-              </Popup>
-            </Marker>
-          );
-        })}
-      </MapContainer>
-
       <Alert className="mb-6 w-full max-w-7xl">
         <FiInfo />
         <AlertTitle>Atenção</AlertTitle>
